# Extend and speedup add_update_test_repo test