- Related ticket: https://openqa.suse.de/tests/3968952#step/add_update_test_repo/88
- Verification run: 
https://openqa.suse.de/tests/3831723
https://openqa.suse.de/tests/3971757
https://openqa.suse.de/tests/3976219 15sp1 before (761 s)
https://openqa.suse.de/tests/3976722 15sp1 after (523 s)
https://openqa.suse.de/tests/3975997 12sp5 before (404 s)
https://openqa.suse.de/tests/3976757 12sp5 after (266 s)
https://openqa.suse.de/tests/3978665 s390x with today network/SCC issues
https://openqa.suse.de/tests/3978678 x86_64 with today network/SCC issues